components/core/requirements.txt
langchain~=0.3.24
langchain-openai~=0.3.14
langchain-community~=0.3.22
langgraph~=0.4.1
allure-pytest~=2.14.1
redis~=6.0.0
tang-yuan-mlops-sdk
langchain-mcp-adapters
python-dotenv~=1.1.0
python-docx~=0.8.11
requests~=2.31.0